+++ rio_tiler_pds/sentinel/aws/sentinel2.py
@@ -3,7 +3,7 @@
 import json
 import re
 from collections import OrderedDict
-from typing import Dict, Tuple, Type
+from typing import Any, Dict, Tuple, Type, Union
 
 import attr
 from morecantile import TileMatrixSet
@@ -13,7 +13,7 @@ from rasterio.warp import transform_geom
 
 from rio_tiler import constants
 from rio_tiler.errors import InvalidBandName
-from rio_tiler.io import BaseReader, COGReader, MultiBandReader
+from rio_tiler.io import COGReader, MultiBandReader
 
 from ... import get_object
 from ..utils import s2_sceneid_parser
@@ -56,7 +56,7 @@ class S2L1CReader(MultiBandReader):
     """
 
     sceneid: str = attr.ib()
-    reader: Type[BaseReader] = attr.ib(default=COGReader)
+    reader: Type[COGReader] = attr.ib(default=COGReader)
     reader_options: Dict = attr.ib(default={"nodata": 0})
     tms: TileMatrixSet = attr.ib(default=constants.WEB_MERCATOR_TMS)
     minzoom: int = attr.ib(default=8)
@@ -192,8 +192,8 @@ class S2L2AReader(S2L1CReader):
 
 
 @attr.s
-class S2COGReader(MultiBandReader):
-    """AWS Public Dataset Sentinel 2 COGS reader.
+class S2L2ACOGReader(MultiBandReader):
+    """AWS Public Dataset Sentinel 2 L2A COGS reader.
 
     Args:
         sceneid (str): Sentinel-2 sceneid.
@@ -206,13 +206,13 @@ class S2COGReader(MultiBandReader):
         stac_item (dict): sentinel 2 COG STAC item content.
 
     Examples:
-        >>> with S2COGReader('S2A_29RKH_20200219_0_L2A') as scene:
+        >>> with S2L2ACOGReader('S2A_29RKH_20200219_0_L2A') as scene:
                 print(scene.bounds)
 
     """
 
     sceneid: str = attr.ib()
-    reader: Type[BaseReader] = attr.ib(default=COGReader)
+    reader: Type[COGReader] = attr.ib(default=COGReader)
     reader_options: Dict = attr.ib(factory=dict)
     tms: TileMatrixSet = attr.ib(default=constants.WEB_MERCATOR_TMS)
     minzoom: int = attr.ib(default=8)
@@ -250,3 +250,25 @@ class S2COGReader(MultiBandReader):
 
         prefix = self._prefix.format(**self.scene_params)
         return f"{self._scheme}://{self._hostname}/{prefix}/{band}.tif"
+
+
+def S2COGReader(sceneid: str, **kwargs: Any) -> S2L2ACOGReader:
+    """Sentinel-2 JPEG2000 readers."""
+    scene_params = s2_sceneid_parser(sceneid)
+    level = scene_params["processingLevel"]
+    if level == "L2A":
+        return S2L2ACOGReader(sceneid, **kwargs)
+    else:
+        raise Exception(f"{level} is not supported")
+
+
+def S2JP2Reader(sceneid: str, **kwargs: Any) -> Union[S2L2AReader, S2L1CReader]:
+    """Sentinel-2 COG readers."""
+    scene_params = s2_sceneid_parser(sceneid)
+    level = scene_params["processingLevel"]
+    if level == "L2A":
+        return S2L2AReader(sceneid, **kwargs)
+    elif level == "L1C":
+        return S2L1CReader(sceneid, **kwargs)
+    else:
+        raise Exception(f"{level} is not supported")